Move resolveSelect mapping from registry to store
